+++ babel.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  presets: ['@babel/preset-typescript'],
-  plugins: ['@babel/plugin-transform-modules-commonjs'],
-};