+++ example/iris/index.js
@@ -1,3 +1,0 @@
-/* eslint-disable */
-require = require('esm')(module);
-module.exports = require('./iris');